requirements.txt
numpy
scikit-learn
pillow
joblib
streamlit==1.29.0
streamlit-drawable-canvas==0.9.3
